requirements.txt
ck==2.6.3
cmake==3.27.4.1
lit==16.0.6
matplotlib==3.7.3
networkx==3.1
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.8
py3nvml==0.2.7
pymoo==0.6.0.1
pynvml==11.5.0
pytorch-lightning==1.5.6
PyYAML==6.0
scikit-learn==1.3.2
scikit-opt==0.6.6
scipy==1.10.1
seaborn==0.12.2
smac==2.0.2
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3
torchmetrics
torchsummary
tensorboard==2.12.0
tqdm==4.66.1
typing_extensions
urllib3==1.26.16
yarl==1.9.2
flask
requests
cryptography
netifaces
logzero